refactor: reorganize test structure and consolidate edge cases
