+++ __tests__/app/manifest.test.ts
